+++ internal/config/compat.go
@@ -127,6 +127,9 @@ func (c *Config) ApplyJavaCompatFromEnv() error {
 	if err = applyBoolEnv("MEMORY_SERVICE_ENCRYPTION_LEGACY_PLAIN_READ_ENABLED", &c.EncryptionLegacyPlainReadEnabled); err != nil {
 		return err
 	}
+	if err = applyBoolEnv("MEMORY_SERVICE_ENCRYPTION_LEGACY_BYTE_V1_READ_ENABLED", &c.EncryptionLegacyByteV1ReadEnabled); err != nil {
+		return err
+	}
 	if err = applyBoolEnv("MEMORY_SERVICE_ENCRYPTION_LEGACY_STREAM_V2_READ_ENABLED", &c.EncryptionLegacyStreamV2ReadEnabled); err != nil {
 		return err
 	}